Possibility to build a Hessian for a subsystem in phonopy-FHI-aims
Support for the FHI-aims 'constrain_relaxation .true.' tag added. Finding this tag after an atom in 'geometry.in', phonopy-FHI-aims will exclude corresponding atom from the Hessian. It is useful for weakly bonded subsystems to calculate phonons only for one part, but with respect to the presence of another (e.g. vibrations of physically adsorbed molecules).