Fix bands for LO-TO in low-D materials
It is important in 1-2D systems the limit q->0,
and the value at which q is evaluated is important,
as there is e.g. in 2D a linear trend towards 0.
When q is then close to 0, the value of q is normalized
to be close to 0 with an hard-coded threshold.
This is still higher than an other hard-coded threshold
for q->0 in the qpoints module, which is set to 1e-5
on all the components.