Allow for `pip install phonopy[plot]` to get plotting dependencies
Right now, if one wants to use the band structure plotting utilities, seekpath needs to be manually installed. To help with this and to aid in CI, it would be convenient to support a shorthand of `pip install phonopy[plot]` (or similar) to install the optional dependencies for plotting. This would also allow phonopy to restrict versions of `seekpath` if incompatibilities ever arose.